+++ packages/blocks/cta-section/src/frontend/index.ts
@@ -1,0 +1,5 @@
+export { CtaSectionPure as Client } from './CtaSection.client';
+export { CtaSection as Server } from './CtaSection.server';
+export { CtaSectionRenderer as Renderer } from './CtaSection.renderer';
+
+export * as Types from './CtaSection.types';