+++ packages/blocks/feature-section-grid/src/frontend/index.ts
@@ -1,0 +1,5 @@
+export { FeatureSectionGridPure as Client } from './FeatureSectionGrid.client';
+export { FeatureSectionGrid as Server } from './FeatureSectionGrid.server';
+export { FeatureSectionGridRenderer as Renderer } from './FeatureSectionGrid.renderer';
+
+export * as Types from './FeatureSectionGrid.types';