+++ packages/blocks/hero-section/src/frontend/index.ts
@@ -1,0 +1,5 @@
+export { HeroSectionPure as Client } from './HeroSection.client';
+export { HeroSection as Server } from './HeroSection.server';
+export { HeroSectionRenderer as Renderer } from './HeroSection.renderer';
+
+export * as Types from './HeroSection.types';