+++ packages/blocks/order-details/src/frontend/OrderDetails.client.tsx
@@ -11,6 +11,10 @@ import { Orders } from '@o2s/framework/modules';
 
 import { cn } from '@o2s/ui/lib/utils';
 
+import { toast } from '@o2s/ui/hooks/use-toast';
+
+import { useGlobalContext } from '@o2s/ui/providers/GlobalProvider';
+
 import { ActionList } from '@o2s/ui/components/ActionList';
 import { InfoCard } from '@o2s/ui/components/Cards/InfoCard';
 import { DynamicIcon } from '@o2s/ui/components/DynamicIcon';
@@ -99,6 +103,7 @@ export const OrderDetailsPure: React.FC<Readonly<OrderDetailsPureProps>> = ({
     ...component
 }) => {
     const { Link: LinkComponent } = createNavigation(routing);
+    const { labels } = useGlobalContext();
 
     const initialFilters: Request.GetOrderDetailsBlockQuery = {
         id: component.id,
@@ -114,34 +119,50 @@ export const OrderDetailsPure: React.FC<Readonly<OrderDetailsPureProps>> = ({
 
     const [isPending, startTransition] = useTransition();
 
-    const handleFilter = (data: Partial<any>) => {
+    const handleFilter = (data: Partial<Request.GetOrderDetailsBlockQuery>) => {
         startTransition(async () => {
-            const newFilters = { ...filters, ...data };
-            const newData = await sdk.blocks.getOrderDetails(
-                {
-                    id: orderId,
-                },
-                newFilters,
-                { 'x-locale': locale },
-                accessToken,
-            );
-            setFilters(newFilters);
-            setItems(newData.productList.products.data);
+            try {
+                const newFilters = { ...filters, ...data };
+                const newData = await sdk.blocks.getOrderDetails(
+                    {
+                        id: orderId,
+                    },
+                    newFilters,
+                    { 'x-locale': locale },
+                    accessToken,
+                );
+                setFilters(newFilters);
+                setItems(newData.productList.products.data);
+            } catch (_error) {
+                toast({
+                    variant: 'destructive',
+                    title: labels.errors.requestError.title,
+                    description: labels.errors.requestError.content,
+                });
+            }
         });
     };
 
     const handleReset = () => {
         startTransition(async () => {
-            const newData = await sdk.blocks.getOrderDetails(
-                {
-                    id: orderId,
-                },
-                initialFilters,
-                { 'x-locale': locale },
-                accessToken,
-            );
-            setFilters(initialFilters);
-            setItems(newData.productList.products.data);
+            try {
+                const newData = await sdk.blocks.getOrderDetails(
+                    {
+                        id: orderId,
+                    },
+                    initialFilters,
+                    { 'x-locale': locale },
+                    accessToken,
+                );
+                setFilters(initialFilters);
+                setItems(newData.productList.products.data);
+            } catch (_error) {
+                toast({
+                    variant: 'destructive',
+                    title: labels.errors.requestError.title,
+                    description: labels.errors.requestError.content,
+                });
+            }
         });
     };