+++ packages/blocks/product-list/src/api-harmonization/product-list.client.ts
@@ -1,0 +1,4 @@
+export const URL = '/blocks/product-list';
+
+export * as Model from './product-list.model';
+export * as Request from './product-list.request';