+++ packages/framework/src/modules/cms/models/blocks/bento-grid.model.ts
@@ -1,0 +1,15 @@
+import { Block, Link, Media } from '@/utils/models';
+
+export class BentoGridBlock extends Block.Block {
+    preTitle?: string;
+    title?: string;
+    description?: string;
+    items?: Card[];
+}
+
+export class Card {
+    title?: string;
+    description?: string;
+    image?: Media.Media;
+    link?: Link.Link;
+}