+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @o2s/docs
 
+## 1.5.1
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+
+## 1.5.0
+
+### Minor Changes
+
+- 1653b74: fix: update FilterItem and renderCell components for improved functionality and code clarity
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+
 ## 1.4.0
 
 ### Minor Changes