+++ packages/blocks/service-list/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @o2s/blocks.service-list
 
+## 1.2.0
+
+### Minor Changes
+
+- ec8794c: improved error handling
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+- Updated dependencies [221dc2c]
+- Updated dependencies [db5b381]
+- Updated dependencies [c2d9438]
+  - @o2s/utils.api-harmonization@0.1.3
+  - @o2s/configs.integrations@0.2.1
+  - @o2s/utils.frontend@0.2.1
+  - @o2s/utils.logger@1.1.3
+  - @o2s/ui@1.7.0
+  - @o2s/framework@1.14.0
+
+## 1.1.1
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+- Updated dependencies [2c780d5]
+- Updated dependencies [1653b74]
+- Updated dependencies [c27726a]
+- Updated dependencies [2c780d5]
+- Updated dependencies [0354126]
+- Updated dependencies [1653b74]
+  - @o2s/framework@1.13.0
+  - @o2s/configs.integrations@0.2.0
+  - @o2s/ui@1.6.0
+  - @o2s/utils.frontend@0.2.0
+  - @o2s/utils.api-harmonization@0.1.2
+  - @o2s/utils.logger@1.1.2
+
 ## 1.1.0
 
 ### Minor Changes