+++ packages/blocks/ticket-list/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @o2s/blocks.ticket-list
 
+## 1.4.0
+
+### Minor Changes
+
+- db5b381: fixing various bugs related with blocks and ui components, adding sign-in button in navigation
+- ec8794c: improved error handling
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+- 235f706: changed the default title from h1 to h2 on a block
+- Updated dependencies [221dc2c]
+- Updated dependencies [db5b381]
+- Updated dependencies [c2d9438]
+  - @o2s/utils.api-harmonization@0.1.3
+  - @o2s/configs.integrations@0.2.1
+  - @o2s/utils.frontend@0.2.1
+  - @o2s/utils.logger@1.1.3
+  - @o2s/ui@1.7.0
+  - @o2s/framework@1.14.0
+
+## 1.3.0
+
+### Minor Changes
+
+- 1653b74: fix: update FilterItem and renderCell components for improved functionality and code clarity
+- 2c780d5: conditionally render button in InfoCard for improved layout
+- 1653b74: feat(ui): tile version of data lists
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+- Updated dependencies [2c780d5]
+- Updated dependencies [1653b74]
+- Updated dependencies [c27726a]
+- Updated dependencies [2c780d5]
+- Updated dependencies [0354126]
+- Updated dependencies [1653b74]
+  - @o2s/framework@1.13.0
+  - @o2s/configs.integrations@0.2.0
+  - @o2s/ui@1.6.0
+  - @o2s/utils.frontend@0.2.0
+  - @o2s/utils.api-harmonization@0.1.2
+  - @o2s/utils.logger@1.1.2
+
 ## 1.2.0
 
 ### Minor Changes