+++ packages/utils/api-harmonization/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @o2s/utils.api-harmonization
 
+## 0.1.3
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+- Updated dependencies [db5b381]
+- Updated dependencies [c2d9438]
+    - @o2s/framework@1.14.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+- Updated dependencies [2c780d5]
+- Updated dependencies [1653b74]
+- Updated dependencies [c27726a]
+- Updated dependencies [2c780d5]
+- Updated dependencies [0354126]
+- Updated dependencies [1653b74]
+    - @o2s/framework@1.13.0
+
 ## 0.1.1
 
 ### Patch Changes