+++ packages/utils/logger/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @o2s/utils.logger
 
+## 1.1.3
+
+### Patch Changes
+
+- 221dc2c: fix: added missing Eslint command to list-staged
+
+## 1.1.2
+
+### Patch Changes
+
+- 0354126: reworked turbo dev task to optimize CPU/RAM usage
+
 ## 1.1.1
 
 ### Patch Changes